Added option to export a game from the editor: - Export game data & runtime exe - change rpath on Linux to ensure exported .so can be loaded

+++ engine/engine/render/VulkanRenderer.cpp
@@ -216,7 +216,7 @@ static void imguiCheckVkResult(VkResult err) {
 }
 
 void Carrot::VulkanRenderer::initImGui() {
-    if (!std::filesystem::exists("imgui.ini")) {
+    if (!std::filesystem::exists("imgui.ini") && std::filesystem::exists("default-imgui.ini")) {
         std::filesystem::copy_file("default-imgui.ini", "imgui.ini");
     }
     imGuiBackend.initResources();
